+++ src/services/event_service.js
@@ -56,6 +56,17 @@ export default (easypostClient) =>
     }
 
     /**
+     * Retrieve the next page of Event collection.
+     * @param {Object} events An object containing a list of {@link Event events} and pagination information.
+     * @param {Number} pageSize The number of records to return on each page
+     * @returns {EasyPostObject|Promise<never>} The retrieved {@link EasyPostObject}-based class instance, or a `Promise` that rejects with an error.
+     */
+    static async getNextPage(events, pageSize = null) {
+      const url = 'events';
+      return this._getNextPage(url, events, pageSize);
+    }
+
+    /**
      * Retrieve an {@link Event event} by its ID.
      * See {@link https://www.easypost.com/docs/api/node#retrieve-an-event EasyPost API Documentation} for more information.
      * @param {string} id - The ID of the event to retrieve.