+++ src/services/insurance_service.js
@@ -35,6 +35,17 @@ export default (easypostClient) =>
     }
 
     /**
+     * Retrieve the next page of Insurance collection.
+     * @param {Object} insurances An object containing a list of {@link Insurance insurances} and pagination information.
+     * @param {Number} pageSize The number of records to return on each page
+     * @returns {EasyPostObject|Promise<never>} The retrieved {@link EasyPostObject}-based class instance, or a `Promise` that rejects with an error.
+     */
+    static async getNextPage(insurances, pageSize = null) {
+      const url = 'insurances';
+      return this._getNextPage(url, insurances, pageSize);
+    }
+
+    /**
      * Retrieve an {@link Insurance insurance} record by its ID.
      * See {@link https://www.easypost.com/docs/api/node#retrieve-an-insurance EasyPost API Documentation} for more information.
      * @param {string} id - The ID of the insurance to retrieve.